Have board create all related objects
Pulls the majority of Unity dependent data into GameData, as I can't figure out how to easily associate Tile assets to Components unless via Editor.

GameData is now one of the few object required to be in the scene, all other objects are created at runtime as part of the board.

This gets closer to allowing for multiple boards generated at will.