Create dynamically sized Boards
This splits the Board.Bounds into two seperate rects, one representing the board in World Space, the other in tile space, who's origin is not tied to the World Origin.

BoardManager also all jazzed up, but it is technically a test object just to visually confirm the game is working as expected.