Have Board use BoardData as its source of truth
Previously the Board factory was exploding a BoardData object into its individual fields, and applying directly against Board's properties.

Makes much more sense to just pass the BoardData as an object of Board, and use that as the canoncial data source.

This allows us to remove the BoardFactory class with its overloaded methods in favour of a single static function in board itself that will bootstrap a Board object- becoming consistent with the static function found in like components.

This also has the Board triggering the Background/Border draw on Awake.

* Add Property/Initialization regions
* Add missing summary on public functions
* Board hold a reference to its Border/Background objects