BoardData Persistence and GameData Asset
I had a list of blocks defined in BoardData, and made the ActivePaintTile assign a painted Tile to that list.

This led me back to my dislike of the GameData gameobject in each scene. I realized there is a better way for this. I converted GameData into a ScriptableObject and added a way to instantiate from Editor. The asset that was created can then be moved into Resources folder and loaded at runtime.

Realizing this, I can probably pull apart GameData from one large object containing all the things into individual assets. This is doubly important as the "Only run once" protobuf thing probably shouldn't be in a place that can be `Resources.Load`ed multiple times.

* Convert GameData object to Asset in Resources
* Create Board.Persistence namespace/folder. Will contain all persistence functionality specific to Board
* Painted tiles added to board.gameData.tiles
  * This will allow for persisted tiles for a board
* Massaged boardData constructor to create the Tiles list