Add contextual menu for Creating a New Board
Previously the Board Editor was only capable of creating a Board with a size of (10, 20).

In order to create a dynamically sized board, we need user input.

To facilitate this requirement, the "New Board" button now triggers the visibility of a "contextual" UI panel.

The NewBoardContextController manages all inputs and buttons for this New Board Context menu, and creates a dynamic board.

Unfortunately, once a variable sized board was created, it would be fit off screen- especially if tied to a mobile aspect ratio.

So now the BoardCamera is also reworked to right-size to the Board size and the aspect ratio.

TODO:
* Validate inputs on NewBoardContext menu board size (or convert to sliders w/ min/max values)
* Clean up MenuButtonController, there is some duplication with NewBoardContextController
* Create contextual menus for Save / Load, which will allow further cleanup of MenuButtonController
* Add some way to pad the camera for BoardEditor. The board will take up the whole screen and the UI elements look ugly on top