Add optional padding to BoardCamera size
The goal is to have the Board centered in the Board Editor. This is working, but the UI overlaps, meaning you can't paint certain tiles.

This change adds the ability to pad the size by an arbitrary amount.

Next we'll have to offset the camera, but I want to commit this working change.