Add functional SaveBoard modal menu
This ended up being large change, which makes sense given it brings the serialization effort full circle. With this change applied, the BoardEditor is fully functional.

There's a few issues here and there, and this change does not include cleanup, but this is a confirmed working state to commit.

The UI system is totally event driven. MenuButtonController acts as a primary driver here, triggering when the New/Load/Save modals appear. As those modals change the "ActiveBoard", it will send event back to MenuButtonController. MenuButtonController sends its own "ActiveBoardEvent", acting a central event source for UI features dependent on a board existing.

With the serialization, I pulled the protobuf surrogate initialization into its own Resource asset. It appears to work just existing in Resources by virtue of the RuntimeInit annotation. Neat stuff.
TODO
- [ ] New Board, paint -> Load Board -> New Board. This flow will show the painted items on the original new board. Should be blank canvas.
- [ ] Input validation on TMP_InputFields still needed
- [ ] Modal class parent for the modal menus. Quite a bit of duplicate code that can be cleaned up.
- [ ] Rip out all that generated save file name garbage. Save file name is now generated by user.