Fully working BoardManager (WIP: Requires Cleanup)
This adds a working, if still messy, BoardManager. The requirements for the game is to have multiple Tetris games playable, so we need a way to turn games off and on in a controlled manner.

However, I do not want to strongly tie the implementation of the game to the management system- I want to turn games on and off without knowing or caring about the games specifically.

This led me to design this feature around interfaces. An `interface` defines common functionality across varying objects.

I started with the GameManager - the thing that turns boards on and off. It interacts with objects implementing the interface `IGameController`. The contract defined by this interface is all related to the lifecycle of our minigame (in this case, the only game so far is `TetrisClassic`).

I created a `GameController` class that implements the interface, and have TetrisClass inheriting from GameController. The idea here is that the GameController defines the basic interaction between Manager/Controller, but if there are new game modes down the line that need more specific implementation, we can implement the interface directly.

As the GameManager deals with the lifecycle of a game, it has to deal with Initialization of each game. This means that it is going to be taking pure data of a game and converting that data into the playable game. This requirement means that I also have to have my serializable data objects implementing a `IGameControllerData`.

The Manager only has knowledge of the abstracted interface methods. The data for each game type can vary wildly, but we need to guarantee data that the manager will use during the Initialization process. Therefore I have converted the `TetrisClassicData` from a struct to a class, so that it can implement the IGameControllerData interface, so that the Manager has the data it needs to implement.

I am shocked this is working, to be honest. I more than once felt that I had architected my way into a corner, and that the implementation of the system is overly-complex. That might still be true, but I believe the complexity helps scale with the intended complexity of the requirements- A lifecycle manager for a game session that does not care what that game actually is- just that it started, it ended, and it provided the feedback required to make those decisions.