Clean up GameManagement and fix rando issues
Added summary/comments for GameManagement items.

Removed GameController + GameControllerData. Originally my thought was to have a base class that could provide default behaviour for IGameController objects, but most of that behaviour was one-liner event invocation. The complexity of base class + interface to save a one liner did not bring joy.

Moving around the position of the board in TetrisClassicManagerTest revealed some minor issues with the Camera/GhostPiece- they were defaulting to 0,0,0- which is a fine position to be in until you move the Board far away. Simple fixes to assign the position.

The only things really to do now is:
- [ ] Implement GameUpdate
- [ ] Implement GameStop
- [ ] Implement GameTerminate